py: Add generated bindings for the Node API
This API will eventually replace the manually written greenlight.proto API.
